# Conflicts:
#	VKR.BLL.NET5/CitiesBL.cs
#	VKR.BLL.NET5/MatchScheduleBL.cs
#	VKR.BLL.NET5/PlayerBL.cs
#	VKR.BLL.NET5/PlayerMovesBL.cs
#	VKR.BLL.NET5/PlayerPositionsBL.cs
#	VKR.BLL.NET5/RostersBL.cs
#	VKR.BLL.NET5/SeasonBL.cs
#	VKR.BLL.NET5/StatsBL.cs
#	VKR.EF.DAO/CitiesEFDAO.cs
#	VKR.EF.DAO/MatchScheduleEFDAO.cs
#	VKR.EF.DAO/PlayerEFDAO.cs
#	VKR.EF.DAO/PlayerMovesDAO.cs
#	VKR.EF.DAO/PlayerPositionsEFDAO.cs
#	VKR.EF.DAO/StandingsEFDAO.cs
#	VKR.EF.DAO/StatsEFDAO.cs
#	VKR.EF.Entities/Tables/TeamRating.cs
#	VKR.PL.NET5/AddCityForm.Designer.cs
#	VKR.PL.NET5/AddCityForm.cs
#	VKR.PL.NET5/AddPlayerForm.cs
#	VKR.PL.NET5/ChangePlayerTeamForm.Designer.cs
#	VKR.PL.NET5/ChangePlayerTeamForm.cs
#	VKR.PL.NET5/LineupsForm.cs
#	VKR.PL.NET5/MatchResultsForm.cs
#	VKR.PL.NET5/PlayerStatsForm.Designer.cs
#	VKR.PL.NET5/PlayerStatsForm.cs
#	VKR.PL.NET5/Properties/Resources.resx
#	VKR.PL.NET5/StatsMenuForm.cs

+++ VKR.EF.Entities/Tables/TeamRating.cs
@@ -8,8 +8,8 @@ namespace VKR.EF.Entities
         public string TeamAbbreviation { get; set; }
         public ushort StrikeZoneProbability { get; set; }
         public ushort HitByPitchProbability { get; set; }
-        public byte SwingInStrikeZoneProbability { get; set; }
-        public byte SwingOutsideStrikeZoneProbability { get; set; }
+        public short SwingInStrikeZoneProbability { get; set; }
+        public short SwingOutsideStrikeZoneProbability { get; set; }
         public ushort HittingProbability { get; set; }
         public ushort FoulProbability { get; set; }
         public ushort SingleProbability { get; set; }
